pyproject.toml
[tool.poetry]
name = "drug-extraction"
version = "0.1.0b"
description = "SDK to extract drugs names and classify them from text."
authors = ["Nick Anthony <nanthony007@gmail.com>"]
license = "GPLv3"
include = [
    {path = 'src/drug_extractions/data/drug_dictionary.csv'}
]

[tool.poetry.dependencies]
python = "^3.9"
ipython = "^7.27.0"
"ruamel.yaml" = "^0.17.16"
pydantic = "^1.8.2"

[tool.poetry.dev-dependencies]
flake8 = "^3.9.2"
black = "^21.8b0"
pytest = "^6.2.5"
isort = "^5.9.3"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
